- Currency issue fixed - Login issue fixed

+++ app/controllers/sessions_controller.rb
@@ -11,15 +11,18 @@ class SessionsController < Devise::SessionsController
     sign_in(scope, resource) unless warden.user(scope) == resource
     respond_to do |format|
       @success = true
-      format.js
-      format.html {redirect_to session.has_key?(:return_url) ? session.delete(:return_url) : root_path}
+      if current_user.role? :super_admin
+        format.html {redirect_to admin_root_path}
+      else
+        format.html {redirect_to session.has_key?(:return_url) ? session.delete(:return_url) : root_path}
+      end
     end
   end
 
   def failure
     respond_to do |format|
       @success = false
-      format.js
+      format.html {redirect_to session.has_key?(:return_url) ? session.delete(:return_url) : root_path}
     end
   end
